backend/requirements.txt
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djoser==2.3.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.2.1
python-dotenv==1.0.1
PyYAML==6.0.1